+++ bindings/node/native/src/extraction.rs
@@ -16,7 +16,7 @@ impl From<Error> for neon::result::Throw {
     fn from(err: Error) -> Self {
         let msg = err.0;
         unsafe {
-            neon_runtime::error::throw_error_from_utf8(msg.as_ptr(), msg.len() as i32);
+            neon_runtime::error::throw_error_from_utf8(msg.as_ptr(), msg.len() as i64);
             neon::result::Throw
         }
     }
@@ -43,13 +43,13 @@ where
 /// This trait provides some extraction helpers, and we implement it for CallContext
 /// so that we can easily extract any type that implements FromJsValue from the arguments.
 pub trait Extract {
-    fn extract<T: FromJsValue>(&mut self, pos: i32) -> LibResult<T>;
-    fn extract_opt<T: FromJsValue>(&mut self, pos: i32) -> LibResult<Option<T>>;
-    fn extract_vec<T: FromJsValue>(&mut self, pos: i32) -> LibResult<Vec<T>>;
-    fn extract_vec_opt<T: FromJsValue>(&mut self, pos: i32) -> LibResult<Option<Vec<T>>>;
+    fn extract<T: FromJsValue>(&mut self, pos: i64) -> LibResult<T>;
+    fn extract_opt<T: FromJsValue>(&mut self, pos: i64) -> LibResult<Option<T>>;
+    fn extract_vec<T: FromJsValue>(&mut self, pos: i64) -> LibResult<Vec<T>>;
+    fn extract_vec_opt<T: FromJsValue>(&mut self, pos: i64) -> LibResult<Option<Vec<T>>>;
 }
 impl<'c, T: neon::object::This> Extract for CallContext<'c, T> {
-    fn extract<E: FromJsValue>(&mut self, pos: i32) -> LibResult<E> {
+    fn extract<E: FromJsValue>(&mut self, pos: i64) -> LibResult<E> {
         let val = self
             .argument_opt(pos)
             .ok_or_else(|| Error(format!("Argument {} is missing", pos)))?;
@@ -57,7 +57,7 @@ impl<'c, T: neon::object::This> Extract for CallContext<'c, T> {
         Ok(ext)
     }
 
-    fn extract_opt<E: FromJsValue>(&mut self, pos: i32) -> LibResult<Option<E>> {
+    fn extract_opt<E: FromJsValue>(&mut self, pos: i64) -> LibResult<Option<E>> {
         let val = self.argument_opt(pos);
         match val {
             None => Ok(None),
@@ -75,7 +75,7 @@ impl<'c, T: neon::object::This> Extract for CallContext<'c, T> {
         }
     }
 
-    fn extract_vec<E: FromJsValue>(&mut self, pos: i32) -> LibResult<Vec<E>> {
+    fn extract_vec<E: FromJsValue>(&mut self, pos: i64) -> LibResult<Vec<E>> {
         let vec = self
             .argument_opt(pos)
             .ok_or_else(|| Error(format!("Argument {} is missing", pos)))?
@@ -85,7 +85,7 @@ impl<'c, T: neon::object::This> Extract for CallContext<'c, T> {
         vec.into_iter().map(|v| E::from_value(v, self)).collect()
     }
 
-    fn extract_vec_opt<E: FromJsValue>(&mut self, pos: i32) -> LibResult<Option<Vec<E>>> {
+    fn extract_vec_opt<E: FromJsValue>(&mut self, pos: i64) -> LibResult<Option<Vec<E>>> {
         self.argument_opt(pos)
             .map(|v| {
                 let vec = v.downcast::<JsArray>()?.to_vec(self)?;
