+++ bindings/python/src/encoding.rs
@@ -86,7 +86,7 @@ impl Encoding {
     }
 
     #[getter]
-    fn get_ids(&self) -> Vec<u32> {
+    fn get_ids(&self) -> Vec<u64> {
         self.encoding.get_ids().to_vec()
     }
 
@@ -96,12 +96,12 @@ impl Encoding {
     }
 
     #[getter]
-    fn get_words(&self) -> Vec<Option<u32>> {
+    fn get_words(&self) -> Vec<Option<u64>> {
         self.encoding.get_words().to_vec()
     }
 
     #[getter]
-    fn get_type_ids(&self) -> Vec<u32> {
+    fn get_type_ids(&self) -> Vec<u64> {
         self.encoding.get_type_ids().to_vec()
     }
 
@@ -111,12 +111,12 @@ impl Encoding {
     }
 
     #[getter]
-    fn get_special_tokens_mask(&self) -> Vec<u32> {
+    fn get_special_tokens_mask(&self) -> Vec<u64> {
         self.encoding.get_special_tokens_mask().to_vec()
     }
 
     #[getter]
-    fn get_attention_mask(&self) -> Vec<u32> {
+    fn get_attention_mask(&self) -> Vec<u64> {
         self.encoding.get_attention_mask().to_vec()
     }
 
@@ -130,11 +130,11 @@ impl Encoding {
             .collect()
     }
 
-    fn word_to_tokens(&self, word_index: u32) -> Option<(usize, usize)> {
+    fn word_to_tokens(&self, word_index: u64) -> Option<(usize, usize)> {
         self.encoding.word_to_tokens(word_index)
     }
 
-    fn word_to_chars(&self, word_index: u32) -> Option<Offsets> {
+    fn word_to_chars(&self, word_index: u64) -> Option<Offsets> {
         self.encoding.word_to_chars(word_index)
     }
 
@@ -142,7 +142,7 @@ impl Encoding {
         self.encoding.token_to_chars(token_index)
     }
 
-    fn token_to_word(&self, token_index: usize) -> Option<u32> {
+    fn token_to_word(&self, token_index: usize) -> Option<u64> {
         self.encoding.token_to_word(token_index)
     }
 
@@ -150,7 +150,7 @@ impl Encoding {
         self.encoding.char_to_token(char_pos)
     }
 
-    fn char_to_word(&self, char_pos: usize) -> Option<u32> {
+    fn char_to_word(&self, char_pos: usize) -> Option<u64> {
         self.encoding.char_to_word(char_pos)
     }
