+++ bindings/python/src/tokenizer.rs
@@ -365,7 +365,7 @@ impl Tokenizer {
     }
 
     #[args(with_added_tokens = true)]
-    fn get_vocab(&self, with_added_tokens: bool) -> PyResult<HashMap<String, u32>> {
+    fn get_vocab(&self, with_added_tokens: bool) -> PyResult<HashMap<String, u64>> {
         Ok(self.tokenizer.get_vocab(with_added_tokens))
     }
 
@@ -591,7 +591,7 @@ impl Tokenizer {
         .into()
     }
 
-    fn decode(&self, ids: Vec<u32>, skip_special_tokens: Option<bool>) -> PyResult<String> {
+    fn decode(&self, ids: Vec<u64>, skip_special_tokens: Option<bool>) -> PyResult<String> {
         ToPyResult(
             self.tokenizer
                 .decode(ids, skip_special_tokens.unwrap_or(true)),
@@ -601,7 +601,7 @@ impl Tokenizer {
 
     fn decode_batch(
         &self,
-        sentences: Vec<Vec<u32>>,
+        sentences: Vec<Vec<u64>>,
         skip_special_tokens: Option<bool>,
     ) -> PyResult<Vec<String>> {
         ToPyResult(
@@ -611,11 +611,11 @@ impl Tokenizer {
         .into()
     }
 
-    fn token_to_id(&self, token: &str) -> Option<u32> {
+    fn token_to_id(&self, token: &str) -> Option<u64> {
         self.tokenizer.token_to_id(token)
     }
 
-    fn id_to_token(&self, id: u32) -> Option<&str> {
+    fn id_to_token(&self, id: u64) -> Option<&str> {
         self.tokenizer.id_to_token(id)
     }
